+++ apps/management-console/scratch.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const prop = 'div';
-const Tag = prop as any;
-const el = <Tag id="1">hello</Tag>;